requirements.txt
discord.py==1.6.0
mcrcon==0.6.0
